Apply keyfn before prefix in WithPrefix
To ensure that key functions up the chain can ignore keys by returning
the empty string, apply the key function in WithPrefix before the other
key functions. This keeps with the rest of the key function
implementations of taking the result of another key function and
modifying it, instead of modifying the key and passing it to a key
function.